feat: improve structure of app to follow more Roda conventions
