+++ app/health_check.rb
@@ -46,8 +46,8 @@ module Html2rss
       # @return [Array<String>]
       def errors
         [].tap do |errors|
-          Parallel.each(LocalConfig.feed_names, in_threads: 4) do |feed_name|
-            Html2rss.feed_from_yaml_config(LocalConfig::CONFIG_FILE, feed_name.to_s).to_s
+          Parallel.each(LocalConfig.feed_names) do |feed_name|
+            Html2rss.feed_from_yaml_config(LocalConfig::CONFIG_FILE, feed_name.to_s)
           rescue StandardError => error
             errors << "[#{feed_name}] #{error.class}: #{error.message}"
           end
